requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
defusedxml==0.7.1
Django==4.1.7
django-autoslug==1.9.9
django-cities==0.6.1
django-cities-light==3.9.2
django-cors-headers==3.10.1
django-filter==23.1
django-ipware==4.0.2
django-model-utils==4.3.1
django-rest-framework-nested==0.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
future==0.18.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.14.1
marshmallow-jsonapi==0.24.0
msgpack==1.0.5
numpy==1.24.3
oic==1.5.0
pandas==1.5.0
Pillow==9.5.0
progressbar2==4.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.17
pyjwkest==1.4.2
PyJWT==1.7.1
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-jose==3.3.0
python-utils==3.6.0
pytz==2022.7.1
redis==4.5.4
requests==2.30.0
rsa==4.9
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tqdm==4.65.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
Unidecode==1.3.6
urllib3==2.0.2
zope.interface==6.0
